convert js implementation to ts for components
